cleanup around lang init
